requirements.txt
openai>=1.0.0
elevenlabs>=0.2.26
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
python-dotenv>=0.19.0
Pillow>=9.5.0
requests>=2.28.0
streamlit>=1.0.0
pyyaml>=6.0.1
pydantic>=2.0.0
